requirements.txt
fastapi
uvicorn[standard]
python-jobspy
